BATM-4126 place moved files to correct packages
